Make request and response header names lower-case
This change introduces a Headers type so as to break backwards
compatibility and trigger compiler errors if the Request.headers()
or Response.headers() methods are called expecting a Map as the
return value.

Header names are lower-cased, to simplify case comparisons and be
compatible with HTTP/2. A part of the reasoning around making
this simplistic change is that lower-casing is very cheap for
a string that is already lower-case, and thus doing the lower-
casing always will only lead to a performance hit the first time
a header is processed. Headers that are propagated internally
will always be lower case when handled by downstream services.