don't handle ERROR_DISPATCH requests from Jetty
If, for instance, an asynchronous request times out (configured
via org.eclipse.jetty.server.HttpChannelState.DEFAULT_TIMEOUT),
Jetty will eventually retry the request. We didn't handle that
case. This change means we check if Jetty has told us there was
a problem with the request, and if so, use that information to
return an error response