Preserve case in com.spotify.apollo.Headers
We own an application that gets requests directly from devices in the wild (that can't be upgraded).  These devices talk a great variety of different protocols (hm1, hm2, mercury multiget, http1, http2, ...)

We eventually want to upgrade to the coming version of Apollo but with the current API it is not possible to respond with headers that have a certain case.

This change makes Apollo preserve the original case of all headers, with the trade-off that they are ordered alphabetically.  This was deemed okay since after a discussion and some investigation we concluded that the order of headers only matters when repeated headers are supported, which Apollo does not do.

Another benefit is that the backing map of the headers is immutable now, which prevents surprising behavior if it accidentally is modified.