Updated WithReasonPhrase.create to handle nulls
If the reasonPhrase being passed into this method is null, the `.replaceFrom` method will throw an NPE.
This can occur easily when doing the following:

```
Status.INTERNAL_SERVER_ERROR
          .withReasonPhrase(ex.getMessage()); // Where the message is nullable
```

This may not be the correct approach. But I have noticed this in our services, and I am open to suggestions.