Attempt to read payload even when contentLength = -1
When sending a request with the header Transfer-Encoding: chunked,
Apollo is concluding that no payload is being sent due to the fact that
getting the content length of such request may return -1 (unknown length).
Such header indicates that the payload is being sent as a series of chunks.

This change tries to read the streamed payload in such a case before
concluding that it is empty.