[metrics] change default histogram reservoir
This changes the default histogram reservoir to a [ReservoirWithTtl](https://github.com/spotify/semantic-metrics/tree/9b51f4f6febe7ca251c9410592324645bdb87d6a#histogram-with-ttl).

This eliminates the impact of latencies like p99 being stuck when instances are drained of traffic or the request throughput of a service or endpoint are low.

This is reservoir has already been battle tested internally by being used in Hermes for server metrics.