Add JUnit 5 service helper extension
The added extension mirrors ServiceHelper rule API and allows to access
the created helper via a getter or an injected method parameter.

Declarative registration of this extension is not allowed and throws an
exception. This is because there is no good way to configure the extension
and specify AppInit or ServiceHelper configuration.