Remove Jetty HTTP service and examples
These are not much used in Spotify. The examples are very old and don't
necessarily represent the usage of Apollo today. The Jetty service and
the related HTTP assembly were good examples of usage of Apollo Open
Source, but nowadays, not having been updated in a while, they risk
creating more confusion than value.