upgrade okio to 2.7.0
A number of internal Spotify libraries related to distributed tracing
are introducing a requirement on okio 2.x and okhttp3 4.x.

Between 1.x and 2.x, okio was rewritten from Java to Kotlin, and
promises to be API compatible with 1.x:
https://square.github.io/okio/changelog/#version-200-rc1

Since Apollo uses okio, this upgrades Apollo to the 2.x line of okio as
a way to verify that there are no problems within Apollo before we start
to upgrade Apollo-using services.