Find SpringBootApplication-annotated class
So that the component scan works as expected. This changes so that the Apollo Spring config tries to find an `@SpringBootApplication`-annotated class and add that as a source to the `SpringApplicationBuilder`.

This makes the component scan work more as expected in that it scans the packages under the found class, which is how Spring Boot normally works.

The main drawback is that the `main` class needs to be annotated with `@SpringBootApplication`, but on the other hand this makes the application a more by-the-book Spring Boot app.

On the other hand this could be a feature: if we find an app annotated with `@SpringBootApplication` we auto-activate the Spring support, only skipping that if `apollo.spring.enabled` is explicitly set to `false`