Fix japicmp errors
The 0.6.1 version of the plugin failed with a ClassNotFoundError on optional dependencies, which the 0.15.3 version does not.

However, it instead failed by throwing an error about a new default method added to the com.spotify.apollo.core.Service interface, which is a feature that was introduced after 0.6.1.

The "overrideCompatibilityChangeParameters" section was added to fix this error.